all video or stream processing scripts refactored
